src/requirements.txt
amqp==5.2.0
billiard==4.2.0
blinker==1.7.0
celery==5.3.5
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
confluent-kafka==2.3.0
cryptography==41.0.5
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.4
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
prompt-toolkit==3.0.41
protobuf==4.21.12
pycparser==2.21
PyMySQL==1.1.0
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.10
Werkzeug==3.0.1
